Publication Ready
The Framing Task is up and running. There are two versions available,
one with the original colour scheme (original) and one colourblind
friendly version (colourblind). It also includes the original
instructions of the task as a pdf in the instructions folder and an
Ipython notebook to easily change the directory of the images for the
trial handlers.

All that is left to do is to verify that both versions of the task run
smoothly and that the data is saved appropriately.