Debugged task
Both version of the task are now running and store all of the relevant
data. An appropriate reference to the Frames, biases and rational
decision-making study has been added at the beginning of the code, all
of the instruction text is of the same size, and wrap width is uniform
throughout the experiment.